+++ tools/grapemcrl2/grapeapp/events/event_referencestate.cpp
@@ -96,8 +96,8 @@ grape_event_remove_reference_state::grape_event_remove_reference_state( grape_fr
   if ( p_normal )
   {
     // Create remove event for transitions that have the state as beginstate.
-    int count = p_ref_state->count_transition_beginstate();
-    for (int i = 0; i < count; ++i)
+    size_t count = p_ref_state->count_transition_beginstate();
+    for (size_t i = 0; i < count; ++i)
     {
       transition* trans_ptr = p_ref_state->get_transition_beginstate( i );
       nonterminating_transition* ntt_ptr = dynamic_cast<nonterminating_transition*> ( trans_ptr );
@@ -117,7 +117,7 @@ grape_event_remove_reference_state::grape_event_remove_reference_state( grape_fr
 
     // Create remove event for nonterminating transitions that have the state as endstate.
     count = p_ref_state->count_transition_endstate();
-    for ( int i = 0; i < count; ++i )
+    for ( size_t i = 0; i < count; ++i )
     {
       nonterminating_transition* ntt_ptr = p_ref_state->get_transition_endstate( i );
       grape_event_remove_nonterminating_transition* event = new grape_event_remove_nonterminating_transition( m_main_frame, ntt_ptr, p_proc_dia_ptr );
@@ -126,7 +126,7 @@ grape_event_remove_reference_state::grape_event_remove_reference_state( grape_fr
 
     // Backup initial designators that designate this state.
     count = p_ref_state->count_initial_designator();
-    for ( int i = 0; i < count; ++i )
+    for ( size_t i = 0; i < count; ++i )
     {
       initial_designator* init_ptr = p_ref_state->get_initial_designator( i );
       grape_event_remove_initial_designator* event = new grape_event_remove_initial_designator( m_main_frame, init_ptr, p_proc_dia_ptr );
